requirements.txt
Pillow
google-generativeai
PyPDF2
streamlit
python-docx
langchain
langchain-core
langchain-google-genai
streamlit-webrtc
whisper